+++ lib/object/is.ts
@@ -3,7 +3,7 @@
  *
  * @param {unknown} val - Value to verify
  */
-function isObject <T extends Record<string, any>> (val:T|unknown):val is T {
+function isObject <T extends Record<string, unknown> = Record<string, unknown>> (val:T|unknown):val is T {
     return Object.prototype.toString.call(val) === '[object Object]';
 }