+++ lib/object/isNotEmpty.ts
@@ -3,9 +3,9 @@
  *
  * @param {unknown} val - Value to verify
  */
-function isNotEmptyObject <T extends Record<string,any>> (val:T|unknown):val is T {
+function isNotEmptyObject <T extends Record<string,unknown> = Record<string, unknown>> (val:T|unknown):val is T {
     if (Object.prototype.toString.call(val) !== '[object Object]') return false;
-    for (const _ in val as Record<string, any>) {
+    for (const _ in val as Record<string, unknown>) {
         return true;
     }
     return false;